+++ parser/param.go
@@ -2,10 +2,11 @@ package parser
 
 type Param struct {
 	Variable
-	pointer bool
+	pointer  bool
+	optional bool
 }
 
-func NewParam(name string, valueType ValueType, layer int, public bool, pointer bool) Param {
+func NewParam(name string, valueType ValueType, layer int, public bool, pointer bool, optional bool) Param {
 	return Param{
 		Variable: Variable{
 			name,
@@ -13,10 +14,15 @@ func NewParam(name string, valueType ValueType, layer int, public bool, pointer 
 			layer,
 			public,
 		},
-		pointer: pointer,
+		pointer:  pointer,
+		optional: optional,
 	}
 }
 
 func (p Param) Pointer() bool {
 	return p.pointer
+}
+
+func (p Param) Optional() bool {
+	return p.optional
 }